+++ test/ex_correios/calculator/request/response_test.exs
@@ -48,6 +48,32 @@ defmodule ExCorreios.Calculator.Request.ResponseTest do
       assert Response.process(response) == expected_response
     end
 
+    test "returns a processed response with the value with the correct format" do
+      response = %HTTPoison.Response{
+        body: File.read!("#{@fixtures_path}/success_high_value_response.xml"),
+        status_code: 200
+      }
+
+      assert Response.process({:ok, response}) ==
+               {:ok,
+                %{
+                  deadline: 5,
+                  declared_value: 0.0,
+                  error: nil,
+                  error_code: "0",
+                  error_message: "",
+                  home_delivery: "S",
+                  manually_entered_value: 0.0,
+                  notes: "",
+                  receiving_alert_value: 0.0,
+                  response_status: "0",
+                  saturday_delivery: "N",
+                  service_code: "04510",
+                  value: 1070.90,
+                  value_without_additionals: 1070.90
+                }}
+    end
+
     test "returns an error response", %{bypass: bypass, calculator_url: calculator_url} do
       expected_response =
         {:ok,